Completed all type hint for evaluator/haspi/*
Fixed all mypy issues that arose as a result. A common issue is that
sampling frequencies are often been treated as integers, whereas they
should be floats.

+++ clarity/enhancer/nalr.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
 
 
 class NALR:
-    def __init__(self, nfir: int, sample_rate: int) -> None:
+    def __init__(self, nfir: int, sample_rate: float) -> None:
         """
         Args:
             nfir: Order of the NAL-R EQ filter and the matching delay
